Its 1 AM and heres what I think I added.. I'm not really thinking too straight lol
So it will convert a string to binary (as of right now MUST be under the AESKey size you made. Will make it split the strings up later). Then XOR that string with the generated key and prints that out. I know this isn't AES encryption yet but I don't want to deal with the matrix yet lol.

Good times