Okay so lots of bug fixes
it ACTUALLY converts it to base64 properly now. I tested it online with a converter to make sure, something I did not do earlier because I was tired and didn't want to.

Also when you generate a key it will always be a multiple of 8, you can't have a key that is not only bytes.

Also, removed those godawful if statements (it did look cool though...) and replaced them with a for loop. I have no idea how I didn't think to do that earlier.