Fixed ANOTHER bug. Now the base64 conversion should work perfectly. PLEASE
also did decryption and added a method that converts binary to a string!

So the decryption just converts the encrypted string to binary, XOR's it, then converts it back to a string.

Things I still need to add are: converting base64 to binary (sounds like literal hell lol) and the actual AES encryption process where i believe there is a matrix and some row shifting goes on to shuffle it even more.