Been awhile, after this I will start working on my text RPG again
So it isn't full on AESEncryption. That is for another time. But it is XOR Encryption with a randomly generated key, which is secure enough for me.

It can now open a file (Although it must be specific, I may fix that soon tbh) and encrypt it to a new file! YAY!